# specs/005-contracts-module/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FlyAccounts
  description: 'FlyAccounts API 6.0.0: global Clients CRUD; server-side search/pagination on clients,
    contracts, people, roles; contracts require clientId on create.'
  version: 6.0.0
servers:
- url: /v1
  description: Versioned API prefix
...
paths:
...
  /roles:
    get:
      operationId: listRoles
      summary: Roles with attached permissions
      description: 'Requires manage_roles. Returns roles as data, including attached existing permissions.
        Creating permission types is not available on this operation.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Roles and their attached permissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesResponse'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - *id001
      - *id002
      - *id003
    post:
      operationId: createRole
      summary: Create a role
      description: 'Requires manage_roles. Name is required and MUST be unique. Existing permissions are
        attached with a later call, not created here.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Created role with attached permissions (none yet).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Another role already uses this name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: duplicate_role_name
                message: A role with that name already exists.
...
  /roles/{roleId}:
    patch:
      operationId: updateRole
      summary: Change a role name or description
      description: 'Requires manage_roles. Name MUST remain unique. Authorization still follows attached
        permissions, not the display name.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Updated role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Another role already uses this name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: duplicate_role_name
                message: A role with that name already exists.
    delete:
      operationId: deleteRole
      summary: Delete an unused role
      description: 'Requires manage_roles. Refused while any User or active Invite is assigned the role.
        Refused when deleting would leave zero signed-in people whose combined permissions include manage_users.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/RoleId'
      responses:
        '204':
          description: Role deleted.
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Role is still assigned, or last access-administration path.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                stillAssigned:
                  value:
                    code: role_still_assigned
                    message: That role is still assigned.
                lastAdmin:
                  value:
                    code: last_admin_required
                    message: At least one person with manage_users must remain.
  /roles/{roleId}/permissions:
    post:
      operationId: attachPermission
      summary: Attach an existing permission to a role
      description: 'Adds the permission without replacing permissions the role already has. Duplicate
        attach is refused. Creating permission types is not available. Requires manage_roles. Refused
        when the change would leave zero signed-in people with manage_users (does not apply to attach
        of a new grant).

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachPermissionRequest'
      responses:
        '200':
          description: Role after the attach.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role or permission not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The role already has this permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: duplicate_permission
                message: That permission is already attached.
  /roles/{roleId}/permissions/{permissionId}:
    delete:
      operationId: detachPermission
      summary: Detach a permission from a role
      description: 'Remaining permissions stay. Every person assigned the role uses the updated combined
        set on their next request. Refused when the change would leave zero signed-in people whose combined
        permissions include manage_users. Requires manage_roles.

        '
      tags:
      - Settings
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/RoleId'
      - $ref: '#/components/parameters/PermissionId'
      responses:
        '200':
          description: Role after the detach.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Combined permissions do not include the required manage_* permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role, permission, or attachment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Would remove the last access-administration grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: last_admin_required
                message: At least one person with manage_users must remain.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: FlyAccounts access JWT from POST /auth/microsoft or POST /auth/refresh.
  parameters:
...
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PermissionId:
      name: permissionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - invalid_token
          - personal_account
          - unknown_tenant
          - unauthorized
          - forbidden
          - pending_access
          - not_found
          - duplicate_assignment
          - last_admin_required
          - duplicate_invite
          - already_present
          - duplicate_role_name
          - role_still_assigned
          - duplicate_permission
          - validation_error
          - entity_context_required
          - invalid_currency
          - invalid_file_type
          - client_in_use
          - duplicate_client_name
        message:
          type: string
...
    Permission:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - permission
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Display label (Permission Name).
        permission:
          type: string
          description: Stable key used in authorization checks. Exact match only.
        description:
          type: string
          description: Optional human-readable explanation.
    Role:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - permissions
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
...
    RolesResponse:
      type: object
      additionalProperties: false
      required:
      - roles
      - page
      - pageSize
      - total
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          enum:
          - 10
          - 25
          - 50
        total:
          type: integer
          minimum: 0
...
    CreateRoleRequest:
      type: object
      additionalProperties: false
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    UpdateRoleRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    AttachPermissionRequest:
      type: object
      additionalProperties: false
      required:
      - permissionId
      properties:
        permissionId:
          type: string
          format: uuid
...